Organized examples into basic, advanced, cluster and stresstest examples, and removed CubeExample and OrientationBoxesExample
