+++ src/test/tests/graphics/scenery/tests/examples/advanced/VolumeExample.kt
@@ -1,15 +1,11 @@
-package graphics.scenery.tests.examples
+package graphics.scenery.tests.examples.advanced
 
 import cleargl.GLVector
 import graphics.scenery.*
 import graphics.scenery.backends.Renderer
-import graphics.scenery.controls.InputHandler
 import graphics.scenery.controls.TrackedStereoGlasses
-import graphics.scenery.controls.behaviours.ArcballCameraControl
-import graphics.scenery.controls.behaviours.FPSCameraControl
 import graphics.scenery.volumes.Volume
 import org.junit.Test
-import org.scijava.ui.behaviour.ClickBehaviour
 import java.io.File
 import java.nio.file.Paths
 import kotlin.concurrent.thread
